requirements.txt
streamlit>=1.31.0
langgraph>=0.2.60
langchain-perplexity>=0.1.0
langchain-core>=0.3.23
pdfplumber>=0.11.0
pydantic>=2.7.4
reportlab>=4.0.9
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.9
